Data Loader completion. Closes #6

+++ src/main/java/com/pg/hppp/repositories/LineRepository.java
@@ -3,5 +3,5 @@ package com.pg.hppp.repositories;
 import com.pg.hppp.model.Line;
 import org.springframework.data.repository.CrudRepository;
 
-public interface LineRepository extends CrudRepository<Line, Long> {
+public interface LineRepository extends CrudRepository<Line, Integer> {
 }
